requirements.txt
aniso8601==9.0.1
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
faunadb==4.1.1
Flask==2.0.2
flask-restplus==0.13.0
flask-restx==0.5.1
future==0.18.2
gunicorn==20.1.0
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.3
iso8601==1.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.3.2
MarkupSafe==2.0.1
pycodestyle==2.8.0
pyrsistent==0.18.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
